+++ baybe/recommenders/pure/nonpredictive/sampling.py
@@ -42,7 +42,6 @@ class RandomRecommender(NonPredictiveRecommender):
             replace=len(disc_candidates) < batch_size,
         )
 
-        cont_random.reset_index(drop=True)
         cont_random.index = disc_random.index
         return pd.concat([disc_random, cont_random], axis=1)